Split native and otel csv marshaling to separate types
